CIWorkflow.yml: multiple bugfixes / cleanup (#868)
## Description

This PR does the following:

1. Adds `alpha` to the cargo release dry run step so that we always bump
the version. This is necessary as `cargo release` attempts to do a dry
run with the current release in the Cargo.toml, which conflicts because
it already exists.

2. Removes the usage and requirements of the crates.io token. This had
multiple issues: (1) Someone with a branch on this repository edit the
CI workflow and use the crates.io token however they wanted. (2) It
caused CI to fail on branches from forks, as we marked the secret as
required and, and secrets are not shared to forks.

3. Removes the editing of some of the rust config files with the
metadata we needed to be able to download crates from azure crates for
running CI.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI

## Integration Instructions

N/A